Revert "PRSD-703: Make ...AreYouSure page inherit from PageWithContentProvider"
This reverts commit c4ae429f0dec557ff8d348c6bbc437d09c3a3337.
