+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/PageWithContentProvider.kt
@@ -5,7 +5,7 @@ import uk.gov.communities.prsdb.webapp.forms.JourneyData
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.FormModel
 import kotlin.reflect.KClass
 
-open class PageWithContentProvider(
+class PageWithContentProvider(
     formModel: KClass<out FormModel>,
     templateName: String,
     content: Map<String, Any>,
