+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/LandlordDeregistrationAreYouSurePageTests.kt
@@ -1,14 +1,20 @@
 package uk.gov.communities.prsdb.webapp.forms.pages
 
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import org.mockito.Mockito.mock
 import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.whenever
 import org.springframework.boot.test.mock.mockito.MockBean
+import org.springframework.http.HttpStatus
 import org.springframework.validation.BindingResult
 import org.springframework.validation.Validator
+import org.springframework.web.server.ResponseStatusException
+import org.springframework.web.servlet.ModelAndView
 import uk.gov.communities.prsdb.webapp.forms.JourneyData
 import uk.gov.communities.prsdb.webapp.forms.PageData
 import uk.gov.communities.prsdb.webapp.forms.steps.DeregisterLandlordStepId
@@ -26,11 +32,48 @@ class LandlordDeregistrationAreYouSurePageTests {
     @BeforeEach
     fun setup() {
         journeyDataService = mock()
+        page = LandlordDeregistrationAreYouSurePage(mutableMapOf(), journeyDataService)
+    }
 
-        page =
-            LandlordDeregistrationAreYouSurePage(mutableMapOf(), journeyDataService) {
-                mock()
-            }
+    @Test
+    fun `enrichModel adds the no properties fieldset heading if the user has no registered properties`() {
+        val journeyData =
+            mutableMapOf(
+                DeregisterLandlordStepId.CheckForUserProperties.urlPathSegment to
+                    mutableMapOf(
+                        USER_HAS_REGISTERED_PROPERTIES_JOURNEY_DATA_KEY to false,
+                    ),
+            ) as JourneyData
+        val modelAndView = ModelAndView()
+
+        page.enrichModel(modelAndView, journeyData)
+
+        assertEquals("forms.areYouSure.landlordDeregistration.noProperties.fieldSetHeading", modelAndView.model["fieldSetHeading"])
+        assertNull(modelAndView.model["fieldSetHint"])
+    }
+
+    @Test
+    fun `enrichModel adds the with properties fieldset heading and hint if the user has registered properties`() {
+        val journeyData =
+            mutableMapOf(
+                DeregisterLandlordStepId.CheckForUserProperties.urlPathSegment to
+                    mutableMapOf(
+                        USER_HAS_REGISTERED_PROPERTIES_JOURNEY_DATA_KEY to true,
+                    ),
+            ) as JourneyData
+        val modelAndView = ModelAndView()
+
+        page.enrichModel(modelAndView, journeyData)
+
+        assertEquals("forms.areYouSure.landlordDeregistration.hasProperties.fieldSetHeading", modelAndView.model["fieldSetHeading"])
+        assertEquals("forms.areYouSure.landlordDeregistration.hasProperties.fieldSetHint", modelAndView.model["fieldSetHint"])
+    }
+
+    @Test
+    fun `enrichModel throws an exception if userHasRegisteredProperties is not found in JourneyData`() {
+        val journeyData = mutableMapOf<String, Any>() as JourneyData
+        val thrownError = assertThrows<ResponseStatusException> { page.enrichModel(ModelAndView(), journeyData) }
+        assertEquals(thrownError.statusCode, HttpStatus.INTERNAL_SERVER_ERROR)
     }
 
     @Test
